% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUFLOWR.R
\docType{package}
\name{TUFLOWR}
\alias{TUFLOWR}
\title{TUFLOWR}
\description{
Helper functions for 'TUFLOW FV' models. Current functionality includes reading in and plotting
output POINTS files and generating initial conditions based on point observations.
}
