% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUFLOWR.R
\docType{package}
\name{TUFLOWR}
\alias{TUFLOWR}
\title{TUFLOWR}
\description{
The SWTools package provides four categories of important functions:
SILO, Aquarius, Source and TUFLOW.
}
\section{TUFLOW functions}{

For plotting POINTS files written by TUFLOW (tested with TUFLOW FV)
}

